Added WordMeaning, rendered the word meanings
getting a response from the dictionary api, want to update results and storing results in setResults; then what doing in WordSearch is rendering the results.

In the results, we store the key word in the h2 and then going through each meaning, one or more; going through each of them and then sending the meaning to the WordMeaning components, which includes multi definitions and prt of spch (noun, verb etc.). In each def, going to display the def and an example.